Switch between localssd and pdssd configuration
If pdSSDSizeGB is set, the nfs server will use pd-ssd as the main
storage, if instead numLocalScratchSSD is set, local ssds will be used
as main storage. One cannot specify arbitrary size for local ssds, can
ony mount 1, 2, 4, 8 ,16, 24 375GB disks depending on the host spec,
right now we allow the users to specify arbitrary number of disks and
depend on google to produce error messages when the configuration is
invalid

+++ cloud/google/nfs_server.py
@@ -1,8 +1,9 @@
-from common import ZonalComputeUrl, GenerateDisk, GenerateBootDisk, GenerateNetworkInterface, INSTALL_DOCKER_CMD, GenerateAirflowVar
+from common import ZonalComputeUrl, GenerateDisk, GenerateBootDisk, GenerateScratchDisk, GenerateNetworkInterface, INSTALL_DOCKER_CMD, GenerateAirflowVar
 from workers import GenerateDockerCommand, GenerateCeleryWorkerCommand, GenerateEnvironVar
 
 
 def GenerateNFSServerStartupScript(context, hostname_manager):
+    nfs_server_param = context.properties["nfsServer"]
     env_variables = GenerateAirflowVar(context, hostname_manager)
 
     docker_env = [f'-e {k}' for k in env_variables]
@@ -39,7 +40,16 @@ http {
     }
 }
 '''
-
+    scratch_nvme_config = '''
+DRIVES=($(lsblk | grep -oE 'nvme[a-z0-9A-Z]*' | cut -d' ' -f1 | awk '{ print "/dev/"$1 }'))
+if [ ${#DRIVES[@]} -ne 0 ]; then
+mdadm --create /dev/md0 --level=0 --force --raid-devices=${#DRIVES[@]} ${DRIVES[@]}
+mkfs.ext4 -F /dev/md0
+mount /dev/md0 /share
+chmod 777 /share
+mkdir -p /share/mariadb
+fi
+'''
     startup_script = f'''
 #!/bin/bash
 set -e
@@ -50,15 +60,17 @@ chmod 777 /var/log/airflow/logs
 
 DEBIAN_FRONTEND=noninteractive apt-get -y -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" dist-upgrade
 {INSTALL_DOCKER_CMD}
+{scratch_nvme_config}
 
 {GenerateEnvironVar(context, env_variables)}
 
 if [ ! -f "/etc/bootstrap_done" ]; then
-
+if ! mountpoint -q "/share"; then
 mkfs.ext4 -F /dev/sdb
 mount /dev/sdb /share
 chmod 777 /share
 mkdir -p /share/mariadb
+fi
 apt-get install nfs-kernel-server nginx -y
 echo "/share 172.31.0.0/16(insecure,rw,async,no_subtree_check)" >> /etc/exports
 echo "ALL: 172.31.0.0/16" >> /etc/hosts.allow
@@ -79,8 +91,10 @@ fi
 
 sysctl -w net.netfilter.nf_conntrack_max=$(awk '/MemAvailable/ {{print int($2/16)}}' /proc/meminfo)
 echo $(awk '/MemAvailable/ {{print int($2/64)}}' /proc/meminfo) > /sys/module/nf_conntrack/parameters/hashsize
+if ! mountpoint -q "/share"; then
 mount /dev/sdb /share
 chmod 777 /share
+fi
 systemctl restart nfs-kernel-server.service
 export INNODB_POOL_SIZE_GB=$(awk '/MemAvailable/ {{print int($2/1024/1024/2)}}' /proc/meminfo)
 export INNODB_LOG_SIZE_GB=$(awk '/MemAvailable/ {{print int($2/1024/1024/8)}}' /proc/meminfo)
@@ -95,12 +109,26 @@ docker run --rm -p 3306:3306 --tmpfs /tmp:rw -v /share/mariadb:/var/lib/mysql --
 def GenerateNFSServer(context, hostname_manager, hostname_nfs_server):
     nfs_server_param = context.properties["nfsServer"]
 
+    if ('pdSSDSizeGB' in nfs_server_param) == ('numLocalScratchSSD' in nfs_server_param):
+        raise ValueError("You must set one and only one of 'pdSSDSizeGB' and 'numLocalScratchSSD'")
+
     startup_script = GenerateNFSServerStartupScript(context, hostname_manager)
 
-    diskType = ZonalComputeUrl(
-        context.env['project'],
-        nfs_server_param['zone'],
-        'diskTypes', 'pd-ssd')
+    disks = [GenerateBootDisk(diskSizeGb=10),]
+
+    if 'pdSSDSizeGB' in nfs_server_param:
+        diskType = ZonalComputeUrl(
+            context.env['project'],
+            nfs_server_param['zone'],
+            'diskTypes', 'pd-ssd')
+        disks.append(GenerateDisk(diskSizeGb=nfs_server_param['pdSSDSizeGB'], diskType=diskType))
+    elif 'numLocalScratchSSD' in nfs_server_param:
+        diskType = ZonalComputeUrl(
+            context.env['project'],
+            nfs_server_param['zone'],
+            'diskTypes', 'local-ssd')
+        disks += GenerateScratchDisk(numDisks=nfs_server_param['numLocalScratchSSD'], diskType=diskType)
+
 
     instance_resource = {
         'zone': nfs_server_param['zone'],
@@ -108,10 +136,7 @@ def GenerateNFSServer(context, hostname_manager, hostname_nfs_server):
                       context.env['project'], nfs_server_param['zone'],
                       'machineTypes', nfs_server_param['machineType']
         ),
-        'disks': [
-                  GenerateBootDisk(diskSizeGb=10),
-                  GenerateDisk(diskSizeGb=nfs_server_param['nfsVolumeSizeGB'], diskType=diskType),
-        ],
+        'disks': disks,
         'labels': {
             'vmrole': 'nfs-server',
             'location': nfs_server_param['zone'],
